# PERF: MultiIndex set and indexing operations