--- a/pandas/_libs/index.pyx
+++ b/pandas/_libs/index.pyx
@@ -720,7 +720,7 @@ cdef class BaseMultiIndexCodesEngine:
         int_keys : 1-dimensional array of dtype uint64 or object
             Integers representing one combination each
         """
-        level_codes = target._recode_for_new_levels(self.levels)
+        level_codes = list(target._recode_for_new_levels(self.levels))
         for i, codes in enumerate(level_codes):
             if self.levels[i].hasnans:
                 na_index = self.levels[i].isna().nonzero()[0][0]