--- a/pandas/core/indexes/multi.py
+++ b/pandas/core/indexes/multi.py
@@ -2399,6 +2399,19 @@ class MultiIndex(Index):
             levels=new_levels, codes=new_codes, names=new_names, verify_integrity=False
         )
 
+    def _recode_for_new_levels(self, new_levels, copy: bool = True) -> list[np.ndarray]:
+        if len(new_levels) != self.nlevels:
+            raise AssertionError(
+                f"Length of new_levels ({len(new_levels)}) "
+                f"must be same as self.nlevels ({self.nlevels})"
+            )
+        return [
+            recode_for_categories(
+                self.codes[i], self.levels[i], new_levels[i], copy=copy
+            )
+            for i in range(self.nlevels)
+        ]
+
     def _get_codes_for_sorting(self) -> list[Categorical]:
         """
         we are categorizing our codes by using the